Update solutions.tex
For Problem 1 under Adders, ALUs, and Bit-Slicing, the question asks for a solution using AND and OR gates only, but the example solutions use XOR gates (these were removed). In Problem 2b, the question says S_3S_2S_1A_0 instead of S_3S_2S_1S_0, this was changed to be correct.